+++ src/Controller/View.h
@@ -12,12 +12,11 @@
 namespace view {
 enum View { MACRO = 0, ANIMATION = 1, FILTER = 2, ANIMATION_SPEED = 3 };
 
-int view_scroll = 0;
-
-int macro_selector = 0;
-int parameter_selector = 0;
-int animation_selector = 2;
-int filter_selector = 0;
+byte view_scroll = 0;
+byte macro_selector = 0;
+byte parameter_selector = 0;
+byte animation_selector = 2;
+byte filter_selector = 0;
 
 void initView();
